update to match new trait hierarchy
