+++ crates/bevy_ecs/src/query/world_query.rs
@@ -150,20 +150,11 @@ pub struct IncludeEntity<D>(pub D);
 /// The only difference is that the `Entity` (which is tracked by all queries internally) is
 /// also included in the output `Item`.
 unsafe impl<D: WorldQuery> WorldQuery for IncludeEntity<D> {
-    /// The item is the same as `D`, but also includes the entity.
-    type Item<'a> = (Entity, D::Item<'a>);
-
     /// The same underlying state is used for `IncludeEntity<D>` as `D` itself.
     type Fetch<'a> = D::Fetch<'a>;
 
     /// The same underlying state is used for `IncludeEntity<D>` as `D` itself.
     type State = D::State;
-
-    /// This function manually implements subtyping for the query items.
-    /// The non-`Entity` part of the query is shrunk exactly how `D` specifies.
-    fn shrink<'wlong: 'wshort, 'wshort>(item: Self::Item<'wlong>) -> Self::Item<'wshort> {
-        (item.0, D::shrink(item.1))
-    }
 
     /// This function manually implements subtyping for the query items.
     /// The query is shrunk exactly how `D` specifies.
@@ -213,23 +204,6 @@ unsafe impl<D: WorldQuery> WorldQuery for IncludeEntity<D> {
     /// Called when constructing a [`QueryLens`](crate::system::QueryLens) or calling [`QueryState::from_builder`](super::QueryState::from_builder)
     fn set_access(state: &mut Self::State, access: &FilteredAccess<ComponentId>) {
         D::set_access(state, access);
-    }
-
-    /// Fetch [`Self::Item`](`WorldQuery::Item`) for either the given `entity` in the current [`Table`],
-    /// or for the given `entity` in the current [`Archetype`]. This must always be called after
-    /// [`WorldQuery::set_table`] with a `table_row` in the range of the current [`Table`] or after
-    /// [`WorldQuery::set_archetype`]  with a `entity` in the current archetype.
-    ///
-    /// # Safety
-    ///
-    /// Must always be called _after_ [`WorldQuery::set_table`] or [`WorldQuery::set_archetype`]. `entity` and
-    /// `table_row` must be in the range of the current table and archetype.
-    unsafe fn fetch<'w>(
-        fetch: &mut Self::Fetch<'w>,
-        entity: Entity,
-        table_row: TableRow,
-    ) -> Self::Item<'w> {
-        (entity, D::fetch(fetch, entity, table_row))
     }
 
     /// Adds any component accesses used by this [`WorldQuery`] to `access`.
